+++ src/plonemeeting/portal/core/tests/test_catalog.py
@@ -1,14 +1,10 @@
 # -*- coding: utf-8 -*-
 
 from datetime import datetime
-
-from Products.PluginIndexes.DateIndex.DateIndex import DateIndex
 from imio.helpers.content import richtextval
 from plone import api
-
-from plonemeeting.portal.core.tests.portal_test_case import (
-    PmPortalDemoFunctionalTestCase,
-)
+from plonemeeting.portal.core.tests.portal_test_case import PmPortalDemoFunctionalTestCase
+from Products.PluginIndexes.DateIndex.DateIndex import DateIndex
 
 
 class TestMeetingWorkflow(PmPortalDemoFunctionalTestCase):