+++ CHANGES.rst
@@ -11,10 +11,17 @@ Changelog
   [odelaere]
 - Handle connection failure properly in institution edit form.
   [odelaere]
+- Amper removing of representatives if they are linked to at least an item.
+  [odelaere]
 - Removed faceted-preview-meeting-items.
   [odelaere]
 - Added DataGridField to manage url parameters.
   [odelaere]
+- Query representatives from iA.Delib to populate vocabularies only when loading the edit form.
+  [odelaere]
+- Fetched representatives from delib are kept if used.
+  [odelaere]
+
 
 1.3.3.2 (2021-08-20)
 --------------------